requirements.txt
Django==1.8.3
django-markdown==0.8.4
django-registration-redux==1.2
